Add solution to smallest substring problem O(N)
It uses two pointers to hold the start and end points of an expanding
windo that only adds a result candidate if all characters in arr exist
between the two pointers. It then shrinks the window until it no longer
satisfies that condition. Each element in the string is touched at most
twice and each element in the char array is touched once when added to
the hashmap giving this a runtime complexity of (2n + m) but since the
char array must be smaller than the string this becomes O(n) runtime and
O(m) space; space could be reduced by using a bit string or boolean array
instead to store characters but that would still be O(n);