Add solution to "Busiest time at the mall" Problem
Solution is bounded by the sorting algorithm since the number of people
cannot be finalized if we have other data entries with the same timestamp.
In addition the current population of the mall is dependant on the previous
timestamps. Thus the solution creates a wrapper object to hold the data and
implements a comparator for it so that data is processed appropriately.
Solution in O(1) space and O(nlog(n)) time